aa
Adjust the blinding of the projectors (Edgar) Pending of help of a friend to make the texture rectangle with photoshop.Me idea is try to make an image with 50% opacity and place it in the zone where the two projectors collides.

Rotate the footprints to see the corners and not the middle (Edgar)(done)

Starting activity (Edgar)

Children are scanned when they are on the footprints (located in the middle)(done)

Children will not see each other since they will be looking at the outer side of the arena.(done)

Once they are scanned, a circle will appear in front of them and after 1 second a group of particles will appear on it (I think the color of the particles need to be different from the Mandala’s Activity - with tonalities of grey and white?).(done)(Pending confirmate color)

The group of particles will not disappear.

When the particles are touched, three rivers will be created and will go to the three circles of the station (the circles will appear at the same time that the particles are touched)(done)

Once the rivers finished, a swarm of particles will appear from the footprints and will go to the new positions (the corner - the station) (this is already done)(redone now every scan zone have their own mechanic)