a
Visual changes:

Texture and color for the footprints (change two of the footprints from the starting sequence to visualize them and choose one) - new (folder): footprint_thiner_1 and footprint_thiner_2 (done)

Change the green color (to activate the footprints): 44E28D(done)

Initial sequence:

The transition swarm needs to start from the footprints (not from the circle) (like in the main activity) (done)

Make the speed of the (transition) swarm slower(done)

The footprints in the stations need to appear once the swarm arrives (like in the transition of the Mandala activity) - verify on Monday (better if we have the chance to see it)(done)

Main sequence (Mandala):

The color of the swarm (for the transition) has to have the color with which the Mandala is going to be painted (the new color, not the previous one to give a perception to the users that the swarm loads the circles with the particles that will appear) (done)

Change the appearance sequence for the particles: Right - Left - Up
(done)

Verify that with two more seconds on the transition the music for all sections matches - ask Rita to have an audio with two more seconds)

Change the orange color of the Mandala (see on Monday in the Lab). Option to visualize (all new):

Blue triangles/circle: 49AEDA

Yellow triangles: FFCD4D

Orange triangles: FF6F4D

Final sequence:

Make the black circle grow slower (done)

Match the black circle with the whole space to not see the Mandala(done)

Add effect sounds both at the beginning and at the end (ask Rita, and if it is not possible look for sounds in free libraries)

Add music (we will have it maximum on the 2nd of May)

Decide how to manage the time when users don't catch particles.

Idea→ accumulate when they didn't catch and activate all the previous if they catch.

-If the finish iterations in that time and doesn't finish triangle fill up when they move. (this will make that the mechanic to have next color to iterate a little crapy, because of the flow system now)